refactor: Refactored LoginMapper to AuthMapper
